220110 change learning rate type

+++ main.py
@@ -12,7 +12,7 @@ if __name__ == '__main__':
     parser.add_argument('--batch_size', type=int, help='batch-size',default=64, required=False)
     parser.add_argument('--d',type=str, help="dataset folder name",default="waseem", required=False)
     parser.add_argument('--cl',type=int, help="number of class",default=3, required=False)
-    parser.add_argument('--lr',type=int, help="learning rate",default=0.01, required=False)
+    parser.add_argument('--lr',type=float, help="learning rate",default=0.01, required=False)
     parser.add_argument('--op',type=str, help="optimizer",default="Adagrad", required=False)
     
     args = parser.parse_args()
